Changed damage of card from int to double

+++ src/main/java/at/if22b208/mtc/dto/card/CardDto.java
@@ -4,5 +4,5 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.util.UUID;
 
-public record CardDto(@JsonProperty("id") UUID uuid, String name, int damage) {
+public record CardDto(@JsonProperty("id") UUID uuid, String name, double damage) {
 }
